Simplify R2DBC configured target handling
