+++ instrumentation/r2dbc-1.0/library/src/main/java/io/opentelemetry/instrumentation/r2dbc/v1_0/internal/DbExecution.java
@@ -16,12 +16,10 @@ import static java.util.stream.Collectors.toList;
 
 import io.opentelemetry.context.Context;
 import io.opentelemetry.instrumentation.api.incubator.semconv.db.internal.DbServerTarget;
-import io.opentelemetry.instrumentation.api.incubator.semconv.db.internal.DbServerTargetBuilder;
 import io.r2dbc.proxy.core.QueryExecutionInfo;
 import io.r2dbc.proxy.core.QueryInfo;
 import io.r2dbc.spi.Connection;
 import io.r2dbc.spi.ConnectionFactoryOptions;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
@@ -33,7 +31,6 @@ import javax.annotation.Nullable;
  * any time.
  */
 public final class DbExecution {
-  private static final int MAX_ENDPOINTS = 5;
   // copied from DbAttributes.DbSystemNameValues
   private static final String POSTGRESQL = "postgresql";
   // copied from DbAttributes.DbSystemNameValues
@@ -125,14 +122,12 @@ public final class DbExecution {
             resolvedDriver,
             resolvedProtocol,
             factoryOptions.hasOption(SSL) && Boolean.TRUE.equals(factoryOptions.getValue(SSL)));
-    ConfiguredServerTarget configuredServerTarget =
-        isUnixDomainSocket(serverAddress)
-            ? ConfiguredServerTarget.from(DbServerTarget.unixSocket(serverAddress))
-            : isServerAddressGroupCandidate(serverAddress)
-                ? buildServerAddressGroup(serverAddress, serverPort, defaultPort)
-                : buildServerTarget(serverAddress, serverPort, defaultPort);
-    this.configuredServerAddress = configuredServerTarget.address;
-    this.configuredServerPort = configuredServerTarget.port;
+    DbServerTarget configuredServerTarget =
+        R2dbcServerTarget.create(serverAddress, serverPort, defaultPort);
+    this.configuredServerAddress =
+        configuredServerTarget == null ? null : configuredServerTarget.getAddress();
+    this.configuredServerPort =
+        configuredServerTarget == null ? null : configuredServerTarget.getPort();
     this.connectionString =
         String.format(
             "%s%s:%s%s",
@@ -223,173 +218,6 @@ public final class DbExecution {
     this.context = context;
   }
 
-  private static boolean isUnixDomainSocket(@Nullable String serverAddress) {
-    return serverAddress != null && serverAddress.startsWith("/");
-  }
-
-  private static boolean isServerAddressGroupCandidate(@Nullable String serverAddress) {
-    return serverAddress != null && serverAddress.indexOf(',') >= 0;
-  }
-
-  private static ConfiguredServerTarget buildServerAddressGroup(
-      @Nullable String serverAddress, @Nullable Integer serverPort, @Nullable Integer defaultPort) {
-    if (serverAddress == null
-        || serverAddress.indexOf('/') >= 0
-        || serverAddress.indexOf('?') >= 0
-        || serverAddress.indexOf('#') >= 0
-        || (serverPort != null && !isValidPort(serverPort))) {
-      return ConfiguredServerTarget.EMPTY;
-    }
-
-    int firstComma = serverAddress.indexOf(',');
-    int userInfoEnd = serverAddress.indexOf('@');
-    if (userInfoEnd > firstComma
-        || (userInfoEnd >= 0 && userInfoEnd != serverAddress.lastIndexOf('@'))) {
-      return ConfiguredServerTarget.EMPTY;
-    }
-
-    String hostList = stripUserInfo(serverAddress);
-    String[] hosts = hostList.split(",", -1);
-    List<ParsedEndpoint> endpoints = new ArrayList<>(hosts.length);
-    for (String host : hosts) {
-      ParsedEndpoint endpoint = parseEndpoint(host, serverPort);
-      if (endpoint == null) {
-        return ConfiguredServerTarget.EMPTY;
-      }
-      endpoints.add(endpoint);
-    }
-    if (defaultPort == null) {
-      return buildUnknownDefaultPortGroup(endpoints);
-    }
-
-    DbServerTargetBuilder builder = DbServerTarget.builder(defaultPort);
-    for (ParsedEndpoint endpoint : endpoints) {
-      builder.addEndpoint(endpoint.host, endpoint.port == null ? -1 : endpoint.port);
-    }
-    return ConfiguredServerTarget.from(builder.build());
-  }
-
-  private static String stripUserInfo(String serverAddress) {
-    int at = serverAddress.lastIndexOf('@');
-    return at < 0 ? serverAddress : serverAddress.substring(at + 1);
-  }
-
-  private static ConfiguredServerTarget buildServerTarget(
-      @Nullable String serverAddress, @Nullable Integer serverPort, @Nullable Integer defaultPort) {
-    if (serverAddress != null
-        && (serverAddress.indexOf('/') >= 0
-            || serverAddress.indexOf('?') >= 0
-            || serverAddress.indexOf('#') >= 0)) {
-      return ConfiguredServerTarget.EMPTY;
-    }
-    if (serverPort != null && !isValidPort(serverPort)) {
-      return ConfiguredServerTarget.EMPTY;
-    }
-    ParsedEndpoint endpoint = parseEndpoint(serverAddress, serverPort);
-    if (endpoint == null) {
-      return ConfiguredServerTarget.EMPTY;
-    }
-    int effectiveDefaultPort = defaultPort != null ? defaultPort : endpoint.port == null ? 1 : -1;
-    return ConfiguredServerTarget.from(
-        DbServerTarget.builder(effectiveDefaultPort)
-            .addEndpoint(endpoint.host, endpoint.port == null ? -1 : endpoint.port)
-            .build());
-  }
-
-  private static ConfiguredServerTarget buildUnknownDefaultPortGroup(
-      List<ParsedEndpoint> endpoints) {
-    for (ParsedEndpoint endpoint : endpoints) {
-      int validationDefaultPort = endpoint.port == null ? 1 : -1;
-      if (DbServerTarget.builder(validationDefaultPort)
-              .addEndpoint(endpoint.host, endpoint.port == null ? -1 : endpoint.port)
-              .build()
-          == null) {
-        return ConfiguredServerTarget.EMPTY;
-      }
-    }
-
-    StringBuilder addressGroup = new StringBuilder();
-    for (int i = 0; i < Math.min(endpoints.size(), MAX_ENDPOINTS); i++) {
-      ParsedEndpoint endpoint = endpoints.get(i);
-      if (addressGroup.length() > 0) {
-        addressGroup.append(',');
-      }
-      appendEndpoint(addressGroup, endpoint);
-    }
-    return new ConfiguredServerTarget(addressGroup.toString(), null);
-  }
-
-  private static void appendEndpoint(StringBuilder result, ParsedEndpoint endpoint) {
-    if (endpoint.port != null && endpoint.host.indexOf(':') >= 0) {
-      result.append('[').append(endpoint.host).append(']');
-    } else {
-      result.append(endpoint.host);
-    }
-    if (endpoint.port != null) {
-      result.append(':').append(endpoint.port);
-    }
-  }
-
-  @Nullable
-  private static ParsedEndpoint parseEndpoint(
-      @Nullable String serverAddress, @Nullable Integer serverPort) {
-    if (serverAddress == null) {
-      return null;
-    }
-    int userInfoEnd = serverAddress.indexOf('@');
-    if (userInfoEnd >= 0 && userInfoEnd != serverAddress.lastIndexOf('@')) {
-      return null;
-    }
-
-    String host = stripUserInfo(serverAddress).trim();
-    if (host.startsWith("[")) {
-      int closingBracket = host.indexOf(']');
-      if (closingBracket < 0 || host.indexOf(']', closingBracket + 1) >= 0) {
-        return null;
-      }
-      String rest = host.substring(closingBracket + 1);
-      Integer port =
-          rest.isEmpty() ? serverPort : parsePort(rest.startsWith(":") ? rest.substring(1) : "");
-      return !rest.isEmpty() && port == null
-          ? null
-          : new ParsedEndpoint(host.substring(1, closingBracket), port);
-    }
-    if (host.indexOf('[') >= 0 || host.indexOf(']') >= 0) {
-      return null;
-    }
-
-    int firstColon = host.indexOf(':');
-    int lastColon = host.lastIndexOf(':');
-    if (firstColon >= 0 && firstColon == lastColon) {
-      Integer port = parsePort(host.substring(firstColon + 1));
-      return port == null ? null : new ParsedEndpoint(host.substring(0, firstColon), port);
-    }
-    return new ParsedEndpoint(host, serverPort);
-  }
-
-  @Nullable
-  private static Integer parsePort(String value) {
-    if (value.isEmpty()) {
-      return null;
-    }
-    int port = 0;
-    for (int i = 0; i < value.length(); i++) {
-      char c = value.charAt(i);
-      if (c < '0' || c > '9') {
-        return null;
-      }
-      port = port * 10 + c - '0';
-      if (port > 65535) {
-        return null;
-      }
-    }
-    return port;
-  }
-
-  private static boolean isValidPort(int port) {
-    return port >= 1 && port <= 65535;
-  }
-
   @Nullable
   private static String resolveDriver(@Nullable String driver, @Nullable String protocol) {
     if (!"pool".equals(driver) || protocol == null) {
@@ -448,33 +276,5 @@ public final class DbExecution {
 
   private static String resolveDbSystemName(@Nullable String driver) {
     return driver != null ? DRIVER_TO_SYSTEM_NAME.getOrDefault(driver, OTHER_SQL) : OTHER_SQL;
-  }
-
-  private static class ParsedEndpoint {
-    private final String host;
-    @Nullable private final Integer port;
-
-    private ParsedEndpoint(String host, @Nullable Integer port) {
-      this.host = host;
-      this.port = port;
-    }
-  }
-
-  private static class ConfiguredServerTarget {
-    private static final ConfiguredServerTarget EMPTY = new ConfiguredServerTarget(null, null);
-
-    @Nullable private final String address;
-    @Nullable private final Integer port;
-
-    private static ConfiguredServerTarget from(@Nullable DbServerTarget target) {
-      return target == null
-          ? EMPTY
-          : new ConfiguredServerTarget(target.getAddress(), target.getPort());
-    }
-
-    private ConfiguredServerTarget(@Nullable String address, @Nullable Integer port) {
-      this.address = address;
-      this.port = port;
-    }
   }
 }
